chore(react): cleans up polyfill logic and provide custom-element polyfills when choosing wc preset
Closes #1250
