+++ packages/web/src/schematics/application/files/app/src/polyfills.ts__tmpl__
@@ -1,3 +1,7 @@
 /**
- * This file contains polyfills loaded on all browsers
- **/
+ * Polyfill stable language features. These imports will be optimized by `@babel/preset-env`.
+ *
+ * See: https://github.com/zloirock/core-js#babel
+ */
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
